write delete modify header apply

+++ src/main/java/com/example/auth/domain/post/post/controller/ApiV1PostController.java
@@ -53,9 +53,36 @@ public class ApiV1PostController {
         );
     }
 
+    record DeleteReqBody(
+            @NotNull
+            Long authorId,
+            @NotBlank
+            @Length(min = 3)
+            String password
+    ) {
+    }
+
     @DeleteMapping("/{id}")
-    public RsData<Void> delete(@PathVariable long id) {
+    public RsData<Void> delete(
+        @PathVariable
+        long id,
+        @RequestHeader Long authorId,
+        @RequestHeader String password
+        ) {
+
+        Member actor = memberService.findById(authorId).get();
+
+        if(!actor.getPassword().equals(password)) {
+            throw new ServiceException("400-1", "비밀번호가 일치하지 않습니다.");
+        }
+
         Post post = postService.getItem(id).get();
+
+        if(post.getAuthor().getId() != authorId) {
+            throw new ServiceException("403-1", "자신이 작성한 글만 삭제 가능합니다.");
+
+        }
+
         postService.delete(post);
 
         return new RsData<>(
@@ -65,13 +92,39 @@ public class ApiV1PostController {
     }
 
 
-    record ModifyReqBody(@NotBlank @Length(min = 3) String title, @NotBlank @Length(min = 3) String content) {
+    record ModifyReqBody(
+            @NotBlank
+            @Length(min = 3)
+            String title,
+            @NotBlank
+            @Length(min = 3)
+            String content
+    ) {
     }
 
     @PutMapping("{id}")
-    public RsData<Void> modify(@PathVariable long id, @RequestBody @Valid ModifyReqBody body) {
+    public RsData<Void> modify(
+            @PathVariable long id,
+            @RequestBody @Valid ModifyReqBody body,
+            @RequestHeader
+            Long authorId,
+            @RequestHeader
+            String password
+    ) {
+
+        Member actor = memberService.findById(authorId).get();
+
+        if(!actor.getPassword().equals(password)) {
+            throw new ServiceException("400-1", "비밀번호가 일치하지 않습니다.");
+        }
 
         Post post = postService.getItem(id).get();
+
+        if(post.getAuthor().getId() != authorId) {
+            throw new ServiceException("403-1", "자신이 작성한 글만 수정 가능합니다.");
+
+        }
+
         postService.modify(post, body.title(), body.content());
         return new RsData<>(
                 "200-1",
@@ -87,21 +140,21 @@ public class ApiV1PostController {
             String title,
             @NotBlank
             @Length(min = 3)
-            String content,
-            @NotNull
-            Long authorId,
-            @NotBlank
-            @Length(min = 3)
-            String password
+            String content
             ) {
     }
 
     @PostMapping
-    public RsData<PostDto> write(@RequestBody @Valid WriteReqBody body) {
+    public RsData<PostDto> write(@RequestBody @Valid WriteReqBody body,
+        @RequestHeader
+        Long authorId,
+        @RequestHeader
+        String password
+    ) {
 
-        Member actor = memberService.findById(body.authorId()).get();
+        Member actor = memberService.findById(authorId).get();
 
-        if(!actor.getPassword().equals(body.password())) {
+        if(!actor.getPassword().equals(password)) {
             throw new ServiceException("400-1", "비밀번호가 일치하지 않습니다.");
         }
